+++ src/main/java/com/dropcube/beans/mocks/Datum.java
@@ -16,7 +16,7 @@ public class Datum {
     private Double precipIntensityError;
     @SerializedName("precipProbability")
     @Expose
-    private Integer precipProbability;
+    private Double precipProbability;
     @SerializedName("precipType")
     @Expose
     private String precipType;
@@ -45,11 +45,11 @@ public class Datum {
         this.precipIntensityError = precipIntensityError;
     }
 
-    public Integer getPrecipProbability() {
+    public Double getPrecipProbability() {
         return precipProbability;
     }
 
-    public void setPrecipProbability(Integer precipProbability) {
+    public void setPrecipProbability(Double precipProbability) {
         this.precipProbability = precipProbability;
     }
